+++ src/unitysvc_services/validator.py
@@ -1,14 +1,23 @@
 """Data validation module for unitysvc_services."""
 
 import json
+import os
 import re
 import tomllib as toml
 from pathlib import Path
 from typing import Any
 from urllib.parse import urlparse
 
+import typer
 from jinja2 import Environment, TemplateSyntaxError
 from jsonschema.validators import Draft7Validator
+from rich.console import Console
+
+
+class DataValidationError(Exception):
+    """Exception raised when data validation fails."""
+
+    pass
 
 
 class DataValidator:
@@ -338,3 +347,169 @@ class DataValidator:
                 results[str(relative_path)] = (is_valid, errors)
 
         return results
+
+    def validate_directory_data(self, directory: Path) -> None:
+        """Validate data files in a directory for consistency.
+
+        Validation rules:
+        1. All service_v1 files in same directory must have unique names
+        2. All listing_v1 files must reference a service name that exists in the same directory
+        3. If service_name is defined in listing_v1, it must match a service in the directory
+
+        Args:
+            directory: Directory containing data files to validate
+
+        Raises:
+            DataValidationError: If validation fails
+        """
+        # Find all JSON and TOML files in the directory (not recursive)
+        data_files: list[Path] = []
+        for pattern in ["*.json", "*.toml"]:
+            data_files.extend(directory.glob(pattern))
+
+        # Load all files and categorize by schema
+        services: dict[str, Path] = {}  # name -> file_path
+        listings: list[tuple[Path, dict[str, Any]]] = []  # list of (file_path, data)
+
+        for file_path in data_files:
+            try:
+                data, load_errors = self.load_data_file(file_path)
+                if load_errors or data is None:
+                    continue
+
+                schema = data.get("schema")
+
+                if schema == "service_v1":
+                    service_name = data.get("name")
+                    if not service_name:
+                        raise DataValidationError(f"Service file {file_path} missing 'name' field")
+
+                    # Check for duplicate service names in same directory
+                    if service_name in services:
+                        raise DataValidationError(
+                            f"Duplicate service name '{service_name}' found in directory {directory}:\n"
+                            f"  - {services[service_name]}\n"
+                            f"  - {file_path}"
+                        )
+
+                    services[service_name] = file_path
+
+                elif schema == "listing_v1":
+                    listings.append((file_path, data))
+
+            except Exception as e:
+                # Skip files that can't be loaded or don't have schema
+                if isinstance(e, DataValidationError):
+                    raise
+                continue
+
+        # Validate listings reference valid services
+        for listing_file, listing_data in listings:
+            service_name = listing_data.get("service_name")
+
+            if service_name:
+                # If service_name is explicitly defined, it must match a service in the directory
+                if service_name not in services:
+                    available_services = ", ".join(services.keys()) if services else "none"
+                    raise DataValidationError(
+                        f"Listing file {listing_file} references service_name '{service_name}' "
+                        f"which does not exist in the same directory.\n"
+                        f"Available services: {available_services}"
+                    )
+            else:
+                # If service_name not defined, there should be exactly one service in the directory
+                if len(services) == 0:
+                    raise DataValidationError(
+                        f"Listing file {listing_file} does not specify 'service_name' "
+                        f"and no service files found in the same directory."
+                    )
+                elif len(services) > 1:
+                    available_services = ", ".join(services.keys())
+                    raise DataValidationError(
+                        f"Listing file {listing_file} does not specify 'service_name' "
+                        f"but multiple services exist in the same directory: {available_services}. "
+                        f"Please add 'service_name' field to the listing to specify which service it belongs to."
+                    )
+
+    def validate_all_service_directories(self, data_dir: Path) -> list[str]:
+        """
+        Validate all service directories in a directory tree.
+
+        Returns a list of validation error messages (empty if all valid).
+        """
+        errors = []
+
+        # Find all directories containing service or listing files
+        directories_to_validate = set()
+
+        for pattern in ["*.json", "*.toml"]:
+            for file_path in data_dir.rglob(pattern):
+                try:
+                    data, load_errors = self.load_data_file(file_path)
+                    if load_errors or data is None:
+                        continue
+
+                    schema = data.get("schema")
+                    if schema in ["service_v1", "listing_v1"]:
+                        directories_to_validate.add(file_path.parent)
+                except Exception:
+                    continue
+
+        # Validate each directory
+        for directory in sorted(directories_to_validate):
+            try:
+                self.validate_directory_data(directory)
+            except DataValidationError as e:
+                errors.append(str(e))
+
+        return errors
+
+
+# CLI command
+app = typer.Typer(help="Validate data files")
+console = Console()
+
+
+@app.command()
+def validate(
+    data_dir: Path | None = typer.Argument(
+        None,
+        help="Directory containing data files to validate (default: ./data or UNITYSVC_DATA_DIR env var)",
+    ),
+):
+    """
+    Validate data consistency in service and listing files.
+
+    Checks:
+    1. Service names are unique within each directory
+    2. Listing files reference valid service names
+    3. Multiple services in a directory require explicit service_name in listings
+    """
+    # Determine data directory
+    if data_dir is None:
+        data_dir_str = os.environ.get("UNITYSVC_DATA_DIR")
+        if data_dir_str:
+            data_dir = Path(data_dir_str)
+        else:
+            data_dir = Path.cwd() / "data"
+
+    if not data_dir.exists():
+        console.print(f"[red]✗[/red] Data directory not found: {data_dir}")
+        raise typer.Exit(1)
+
+    console.print(f"[cyan]Validating data files in:[/cyan] {data_dir}")
+    console.print()
+
+    # Create validator and run validation
+    validator = DataValidator(data_dir, data_dir.parent / "schema")
+    validation_errors = validator.validate_all_service_directories(data_dir)
+
+    if validation_errors:
+        console.print(f"[red]✗ Validation failed with {len(validation_errors)} error(s):[/red]")
+        console.print()
+        for i, error in enumerate(validation_errors, 1):
+            console.print(f"[red]{i}.[/red] {error}")
+            console.print()
+        raise typer.Exit(1)
+    else:
+        console.print("[green]✓ All data files are valid![/green]")
